+++ services/web/server/src/simcore_service_webserver/application.py
@@ -25,10 +25,10 @@ from .invitations import setup_invitations
 from .login.plugin import setup_login
 from .long_running_tasks import setup_long_running_tasks
 from .meta_modeling.plugin import setup_meta_modeling
+from .notifications.rabbitmq import setup_rabbitmq
 from .products import setup_products
 from .projects.plugin import setup_projects
 from .publications import setup_publications
-from .rabbitmq.rabbitmq import setup_rabbitmq
 from .redis import setup_redis
 from .remote_debug import setup_remote_debugging
 from .resource_manager.plugin import setup_resource_manager
