+++ services/web/server/src/simcore_service_webserver/rabbitmq_settings.py
